Initial commit for c#/.net bdatum library
This includes the basic http communication, and also the first versions
of each call.

Still missing some calls and some constructors. This source is based on
the first version that I did for tests and now it is being organized to
be a nice class.

Inside the directory is a doc.txt, that is the usage/library
description.